--- edu/jalc/automobile/component/engine/EngineAssembly.java
@@ -1,4 +1,0 @@
-package edu.jalc.automobile.component.engine;
-
-abstract public class EngineAssembly{
-}
