--- edu/jalc/automobile/component/engine/README.md
@@ -1,1 +1,0 @@
-#The Automobile Engine
