--- edu/jalc/automobile/component/engine/exhaust/Exhaust.java
@@ -1,3 +1,0 @@
-package edu.jalc.automobile.component.engine.exhaust;
-abstract public class Exhaust{
-}
