--- edu/jalc/automobile/component/engine/induction/BoostedInduction.java
@@ -1,4 +1,0 @@
-package edu.jalc.automobile.component.engine.induction;
-abstract public class BoostedInduction{
-
-}
