--- edu/jalc/automobile/component/engine/induction/NaturallyAspiratedInduction.java
@@ -1,6 +1,0 @@
-package edu.jalc.automobile.component.engine.induction;
-public class NaturallyAspiratedInduction{
-  public String toString(){
-    return "naturally aspirated induction";
-  }
-}
