--- edu/jalc/automobile/component/engine/output/Torque.java
@@ -1,22 +1,0 @@
-package edu.jalc.automobile.component.engine.output;
-public class Torque{
-   
-   private double torque;
-   private double rpm;
-   
-   private Torque(){
-      this.torque = this.rpm = 0;
-   }
-   
-   private Torque(double torque, double rpm){
-      this.torque = torque;
-      this.rpm = rpm;
-   }
-   
-   public double getTorque(){
-      return this.torque;
-   }
-   public double getRPM(){
-      return this.rpm;
-   }
-}
