+++ edu/jalc/automobile/onlinebuilder/components/engine/EngineAssembly.java
@@ -1,0 +1,4 @@
+package edu.jalc.automobile.onlinebuilder.components.enginecomponents;
+
+abstract public class EngineAssembly{
+}
