+++ edu/jalc/automobile/onlinebuilder/components/engine/SportEngineAssembly.java
@@ -1,0 +1,7 @@
+package edu.jalc.automobile.onlinebuilder.components.enginecomponents;
+
+import edu.jalc.automobile.parts.engineparts.exhaust.PerformanceExhaust;
+
+abstract public class SportEngineAssembly extends EngineAssembly{
+  private PerformanceExhaust exhaust;
+}
