+++ edu/jalc/automobile/onlinebuilder/components/enginecomponents/EngineAssembly.java
@@ -1,4 +1,4 @@
-package edu.jalc.automobile.component.engine;
+package edu.jalc.automobile.onlinebuilder.components.engine;
 
 abstract public class EngineAssembly{
 }
