+++ edu/jalc/automobile/onlinebuilder/components/enginecomponents/SportEngineAssembly.java
@@ -1,0 +1,7 @@
+package edu.jalc.automobile.onlinebuilder.components.engine;
+
+import edu.jalc.automobile.parts.engineparts.exhaust.PerformanceExhaust;
+
+abstract public class SportEngineAssembly extends EngineAssembly{
+  private PerformanceExhaust exhaust;
+}
