+++ edu/jalc/automobile/parts/engineparts/exhaust/EconomyExhaust.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.exhaust;
+public class EconomyExhaust extends Exhaust{
+  public String toString(){
+    return "economy exhaust";
+  }
+}
