+++ edu/jalc/automobile/parts/exhaust/EconomyExhaust.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.engineparts.exhaust;
+public class EconomyExhaust extends Exhaust{
+  public String toString(){
+    return "economy exhaust";
+  }
+}
