+++ edu/jalc/automobile/parts/exhaust/Exhaust.java
@@ -1,0 +1,3 @@
+package edu.jalc.automobile.parts.engineparts.exhaust;
+abstract public class Exhaust{
+}
