+++ edu/jalc/automobile/parts/exhaust/PerformanceExhaust.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.engineparts.exhaust;
+public class PerformanceExhaust extends Exhaust{
+  public String toString(){
+    return "performance exhaust";
+  }
+}
