+++ edu/jalc/automobile/parts/induction/BoostedInduction.java
@@ -1,0 +1,4 @@
+package edu.jalc.automobile.parts.induction;
+abstract public class BoostedInduction extends Induction{
+
+}
