+++ edu/jalc/automobile/parts/induction/Induction.java
@@ -1,4 +1,4 @@
-package edu.jalc.automobile.component.engine.induction;
+package edu.jalc.automobile.parts.induction;
 abstract public class Induction{
   abstract public String toString();
 }
