+++ edu/jalc/automobile/parts/induction/NaturallyAspiratedInduction.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.induction;
+public class NaturallyAspiratedInduction extends Induction{
+  public String toString(){
+    return "naturally aspirated induction";
+  }
+}
