+++ edu/jalc/automobile/parts/induction/SuperchargedInduction.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.induction;
+public class SuperchargedInduction extends BoostedInduction{
+  public String toString(){
+    return "supercharged induction";
+  }
+}
