+++ edu/jalc/automobile/parts/induction/TurbochargedInduction.java
@@ -1,0 +1,6 @@
+package edu.jalc.automobile.parts.induction;
+public class TurbochargedInduction extends BoostedInduction{
+  public String toString(){
+    return "turbocharged induction";
+  }
+}
